Rename gis-client TypeDescriptor to PortTypeDescriptor
Moved File from North52 implementation to gis-client
